# Diseño/api-docs.json
{"swagger":"2.0","info":{"description":"Servicios REST para capacitaciones ITEHL","version":"1.0","title":"API Capacitaciones ITEHL","termsOfService":"http://codmind.com/terms","contact":{"name":"Juan José Buzón","url":"https://www.linkedin.com/in/jjbs2406/","email":"j2buzon@gmail.com"},"license":{"name":"License of API","url":"API license URL"}},"host":"localhost:8080","tags":[{"name":"curso-rest-controller","description":"Curso Rest Controller"},{"name":"descuento-rest-controller","description":"Descuento Rest Controller"},{"name":"modalidad-rest-controller","description":"Modalidad Rest Controller"}],"paths":{"/capacitaciones/api/cursos":{"get":{"tags":["curso-rest-controller"],"summary":"listCourses","operationId":"listCoursesUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Curso"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["curso-rest-controller"],"summary":"saveCourse","operationId":"saveCourseUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"course","description":"course","required":true,"schema":{"$ref":"#/definitions/Curso"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Curso"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["curso-rest-controller"],"summary":"updateCourse","operationId":"updateCourseUsingPUT","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"course","description":"course","required":true,"schema":{"$ref":"#/definitions/Curso"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Curso"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/capacitaciones/api/cursos/modalidad/{id}":{"get":{"tags":["curso-rest-controller"],"summary":"listCoursesByModality","operationId":"listCoursesByModalityUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Curso"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/capacitaciones/api/cursos/{id}":{"get":{"tags":["curso-rest-controller"],"summary":"findCourseById","operationId":"findCourseByIdUsingGET","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Curso"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/capacitaciones/api/descuentos":{"get":{"tags":["descuento-rest-controller"],"summary":"getDiscounts","operationId":"getDiscountsUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Descuento"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["descuento-rest-controller"],"summary":"saveDiscount","operationId":"saveDiscountUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"discount","description":"discount","required":true,"schema":{"$ref":"#/definitions/Descuento"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Descuento"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/capacitaciones/api/modalidades":{"get":{"tags":["modalidad-rest-controller"],"summary":"getModalities","operationId":"getModalitiesUsingGET","produces":["application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Modalidad"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["modalidad-rest-controller"],"summary":"saveModality","operationId":"saveModalityUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"modality","description":"modality","required":true,"schema":{"$ref":"#/definitions/Modalidad"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Modalidad"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"Curso":{"type":"object","properties":{"costo":{"type":"number","format":"double"},"descripcion":{"type":"string"},"dirigidoA":{"type":"string"},"horas":{"type":"integer","format":"int32"},"id":{"type":"string"},"modalidad":{"$ref":"#/definitions/Modalidad"},"nombre":{"type":"string"}},"title":"Curso"},"Descuento":{"type":"object","properties":{"curso":{"$ref":"#/definitions/Curso"},"fechaFinal":{"type":"string"},"id":{"type":"string"},"modalidad":{"$ref":"#/definitions/Modalidad"},"nombre":{"type":"string"},"pais":{"type":"string"},"valorDescuento":{"type":"number","format":"double"}},"title":"Descuento"},"Modalidad":{"type":"object","properties":{"descripcion":{"type":"string"},"id":{"type":"string"},"nombre":{"type":"string"}},"title":"Modalidad"}}}